Move CLI config commands into seperate file.
