+++ lib/pivotal_markdown/version.rb
@@ -1,3 +1,3 @@
 module PivotalMarkdown
-  VERSION = "0.0.1"
+  VERSION = "0.0.3"
 end
